Comic_Strip_Generator/requirements.txt
streamlit>=1.36
Pillow>=10.3
python-dotenv>=1.0
google-genai>=0.3.0
